% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packagegename-data.R
\docType{data}
\name{demo}
\alias{demo}
\title{A mini dataset}
\format{
We put a mini dataset in ACAM.
}
\source{
Obtained from GEO with the accession number: GSE99701.
}
\description{
A mini dataset
}
\details{
\code{DF_chen} is the dataset Chen obtained from GEO with the accession number: GSE99701. \code{GM_chen} is the species- and tissue-specific marker genes obtained from the database CellMatch, \code{Y_chen.raw} is the true label, and \code{chen_ACAMclusters} is the clustering results obtained from the function \code{ACAM_cluster}.
}
\examples{
data(DF_chen)
str(DF_chen)
}
