% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louvain.R
\name{Louvain}
\alias{Louvain}
\title{Louvain}
\usage{
Louvain(X)
}
\arguments{
\item{X}{The input dataset.}
}
\value{
The clustering results \it{cluster_results}.
}
\description{
The function Louvain assembles the result of five clustering methods used in the function \it{ACAM_cluster}.
}
\details{
This is the description of the function Louvain.

You can use this function to calculate x+1,then return the value of x+1.
}
