% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACAM_annotate.R
\name{ACAM_annotation}
\alias{ACAM_annotation}
\title{ACAM_annotation}
\usage{
ACAM_annotation(
  DF,
  cluster_results,
  gene.markers,
  min_num = 10,
  k.neighbors = 1,
  pca.rank = 50,
  umap.rank = 10
)
}
\arguments{
\item{DF}{The input dataset. Make sure that the dataset is lognormalized. Note that the rows are cells, and the columns are marker genes.}

\item{cluster_results}{The clustering results obtained from the function \it{ACAM_cluster}}

\item{gene.markers}{The species- and tissue-specific marker genes obtained from CellMatch database.}

\item{min_num}{The mininum number which any clusters have a larger size than this number will be considered as representative clusters. Default is 10.}

\item{k.neighbors}{The number of neighbors considered in the final step kNN.}

\item{pca.rank}{The dimension of which the dataset will be reduced into by the method PCA.}

\item{umap.rank}{The dimension of which the dataset will be reduced into by the method umap.}
}
\value{
The annotation results \it{annotation_results}.
}
\description{
The function ACAM_annotation annotates cells by the method ACAM.
}
\details{
This is the description of the function ACAM_annotation.

The function ACAM_annotation annotates cells by the method ACAM.
}
